requirements.txt
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
boto3==1.35.8
botocore==1.35.8
bytecode==0.15.1
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Deprecated==1.2.14
dj-database-url==2.2.0
Django==5.0.7
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==4.4.0
django-storages==1.14.4
django-taggit==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
envier==0.5.2
exceptiongroup==1.2.2
flower==2.0.1
h11==0.14.0
haversine==2.8.1
httpcore==1.0.5
httpx==0.27.2
humanize==4.10.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.4.0
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
kombu==5.4.0
msgpack==1.0.8
opentelemetry-api==1.27.0
packaging==24.1
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.28.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
service-identity==24.1.0
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
standardwebhooks==1.0.0
tomli==2.0.1
tornado==6.4.1
Twisted==24.7.0
txaio==23.1.1
types-Deprecated==1.2.9.20240311
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
xmltodict==0.13.0
zipp==3.20.1
zope.interface==7.0.1
